requirements.txt
tqdm
plotly
chromadb
streamlit
google.generativeai
scikit-learn
llama_index
sec_edgar_downloader
seaborn